Hashed out basic implementation of distance calculation using RAxML.
Similar to alignment, idea is to take a single class that chooses a method
and farms out calls after validation. Beginning to think it might be
better to separate out classes for each separate call and then pass args on
to each? For BioPython-based modules, this could make it more explicit
where the validation occurs. Still probably have to validate the command
though (BioPython classes only validate parameters?).