Started work on internal ScrollSeq objects.
ScrollSeq is intended to represent a sequence in ScrollPy in terms of:

1) Attributes necessary for sorting/output (group and distance)

2) (Optional) a parsed BioPython SeqRecord object
    -allows for dynamic creation of files and for attribute access

Currently only implement __iadd__, which is necessary. Debatable whether
other arithmetic methods need to be added? In case of sorting, might be
easiest to add comparison operators so that the objects themselves can be
sorted in a meaningful way (e.g. by distance).