First steps of implementing logging complete.
Plan is to have three different types of root loggers:

1) Console
2) File
3) RawOutput

Can send logging messages to each of these in the form of small class
instances defining __str__. Additionally, use filters for each to
determine how to handle messages based on context, attributes, and the
user-defined verbosity argument.

Mostly written, but need to decide on exception output and then also
try to test the code somehow.