Working on filtering.
For now, just filtering based on Z-score. Current implementation appears
to work with a single group of sequences and writes both the desired
output sequences as well as the removed sequences to output files.

Next need to check whether it works with more than one file.

Also, probably want to provide an option in the future to write a
separate output table that is just the removed sequences and something
about their metrics to be removed?