Added test code for current filter implementation.
Added code for current implementation of Z-score filtering. Mocks data
so that actual input is not required and to ensure that distribution
allows for outlier detection.

Considering changing old filtering code to be more flexible. Mainly
want:

1) Ability to use a mapping if only a single file is provided (to-do
upstream, will involve creation of a seq_dict to feed in)

2) Ability to apply filtering method to either all pooled sequences
(current mode) or to apply to each group independently (to-do)

3) Filter based not just on length but also on similarity (to-do)

In order to make these changes, will create a second file and stretch
out implementation to use multiple classes.