Starting on ScrollTree implementation.
Started on an implementation of a SrollTree object (similar to
ScrollPy), which can perform all-v-all tree distance comparisons between
a set of leaves and subsequently output those leaves as labels (to a
table) and/or sequences (assuming they were provided in the first
place).

Current implementation provides a wrapping class for terminal leaf nodes
that delegate distance calculation to the wrapped node but store an
internal distance attribute, and can also wrap a ScrollSeq object as
well.

Planning to change SrollPy to take as input a pre-populated seq_dict
(like ScrollTree as currently written), which means mapping and creating
these dicts is to be moved to a util module with common functionality.
Work on this next.