Finished first implementation of mapping.
Started initial class-based mapping implementation. Allows flexible use
of any input files to produce a mapping, but places preference on user
input in order: mapfile > seqfiles > treefile.

Given a mapping, tries to build a seq_dict based on that mapping that
will group required objects for underlying ScrollSaw operations, but
that is flexible enough to support trees and/or sequences.

Need to add custom pairwise alignment code for label/name pairs that are
of unequal lengths and improve error reporting. Importantly, should halt
program execution if any relationships are not 1-to-1.