Added tree file parsing for newick files.
Parsing of trees to be handled exclusively by ETE3 (for now, at least).

Basic tree parsing uses only Newick format, but this shouldn't be a
problem. Tries to read in the most complete format first, and will
eventually log any failures to do so; eventually, a very flexible format
is used to ensure that any actual Newick files will be parsed.

TO-DO:
    -improve testing
    -deal with other files formats(?)