Improved Mapping and started test coverage.
Moved the unique group name function out of Mapping class and changed it
to use a dictionary rather than a set and counter. This function in
general is problematic, and necessitated setting a '_test' flag on the
_parse_infiles() Mapping method to allow for reproducible tests without
spawning multiple independent groups. This might be a target for future
work to change the implementation.

Completed test coverage for the standalone methods and for the Mapping
class using a single sequence file as the input.

TO-DO: Improve test coverage of Mapping class using multiple sequence
files and/or a treefile/mapfile. Ensure that class behaves as expected
and produces the appropriate seq_dict.