Continuing with mapping testing.
Continuing to build up test suite for functions and classes in the
_mapping module.

During testing, decided to change the way the _best_name_match function
(and functions it uses) works. Although the aligned pairs are used to
determine the best match to a set of names, what is really needed as a
return value is the original (unaligned/changed) name to use as the best
matched label. Therefore, the function now keeps track of both original
and aligned values and returns the original instead of the modified one.