Added warning suppression for ETE3 tree parsing.
ETE3 appears to open files using the deprecated 'U' mode. This warning
does not appear during interpreter sessions or in other scripts but
shows up in unittest output.

Wrapped the function in a simple with statement to suppress these
warnings. Should eventually make sure that exceptions are still thrown
properly even with this change.