Working on updating program execution and funtionality.
Made some changes to __main__ and to the SeqWriter class in output. Idea
is to allow for filtering only runs, if desired. Updated test suite for
output classes to work with new mapping model.

Still need to sort out how filtering works when there are no sequences
provided (if it works at all?). Presumably with just a tree file (and
optional mapping file) the goal is to trim long branches, which is what
ScrollTree does already.

Also need to add actual code for filtering only. Sequences are the
obvious output, but would like to also be able to provide a summary.
This could be done by modifying TableWriter class, or by making a
different kind of summary file. Would also be nice to create a summary
file for program runs in general.