Updated mapping to handle alignments.
For TreePlacer, needed to associate alignment records with tree leaves
as well. Updated Mapping object to take alignments as input as well.

If alignments are provided together with infiles for whatever reason
(for example, if the user wants to still create a mapping from starting
sequence files that were then catted and aligned), the mapping itself
can come from the sequence files, but the SeqRecord objects are those in
the alignment. If need be, can write in a method that removes gaps for
writing out unaligned sequences as well.

Note: for now (at least), no option to create a mapping from an
alignment (like a treefile, it would only produce a single group anyway
and so is not inherently useful).