Added bulk of code for tree placing.
Added TreePlacer class and associated methods. Takes a starting mapping,
an alignment, and a list of BioPython SeqRecord objects to place and
iteratively runs Mafft/IQtree to place each sequence in the existing
alignment, run a new tree, and then classify the sequence based on the
new tree.

Other changes:
    -Updated Aligner class to handle adding sequences to an existing
    alignment (BioPython commandline wrapper is not flexible so had to
    use subprocess directly)
    -Added a TreeBuilder class, similar to Aligner and Distance, to
    handle checking arguments and building trees
    -Added utility functions for working with ETE3 trees to the _tree
    util module and also added test code

Need to check through all of this code and thoroughly test it before
moving on to integrating it in __main__