Started testing TreePlacer.
Started test coverage but immediately ran into an issue with original
scheme for temporary sequence files in TreePlacer. Opening a temp file
and writing to it in a context block did not work as the open file had
size=0 within the block. Closing the file would likely have resulted in
it being garbage collected anyway.

Instead, added sequence files to the _get_outpath() function and
allowing them to reside in the same output directory, which can still be
deleted following execution based on user input.