Finished TreePlacer testing with simple input tree.
Added code to both _classify_node and _classify_monophyletic_node
functions in TreePlacer class. For non-monophyletic nodes, all
constituent groups are listed along with support for the first
monophyletic group that is found for each during traversal.

In general, all monophyletic groups are also checked for complete leaf
membership, i.e. whether all members of a group (based on mapping) are
present under a given node.

Next need to consider output files, how best to tie this in with output
from other modes, and integrating all of this into __main__.