Added additional options to __main__.
Added more options to __main__ to handle new and planned analysis modes,
as well as updated main execution section to reflect changes as well.

One question that arises is how to deal with Filter and SeqWriter.
Realized now that filtering provides both filtered sequences as well as
those that remain. Presumably want to write out all those sequences when
'--filter-out' is specified by user.

Work on this next and then begin to work on final analysis mode, iter
over an alignment to select columns. Then work on finishing touches to
ScrollPy for initial version commit:
    -consistent use of config
    -logging
    -arg and input checking
    -improve test code throughout
    -documentation