Fixed bug in logging newlines.
Realized that logging newlines in modules relies on handlers attached to
root logger, but first implementation looked for handlers directly on
the logger passed as argument. Therefore, added a quick check for
attached handlers and a separate call to find the root logger when the
function is called from a deeper module.

Also realized that care needs to be taken when using status_logger ->
any additional console_logger calls in more deeply nested code will
result in the lines being clobbered messily. In general, need to
determine what things to log to console in the first place and how it
correlates with verbosity level.