Added new code to logging module to log exceptions.
Logging exceptions 'as-is' using built-in logging module is probably not
ideal, at least for console output. Instead, implement a few changes
that allow intercepting and formatting the traceback output for both
console and file output.

Actual logging calls in application code may result in changing some of
this implementation in future.