Added more logging calls and output logic.
Added logging calls to the output module. In a few places, added
additional logic for 'no-create' and 'no-clobber' flags. The 'no-create'
flag may only be relevant in __main__ where the output directory is set,
unless the implementation changes at some point to make use of subdirs?

The 'no-clobber' flag is relevant for all possible output files. It may
be possible to factor out a single function to retrieve output file
names, as a similar function appears in multiple places.