Added central function to obtain output filepaths.
A similar function existed throughout ScrollPy in many different classes
to obtain expected output filepaths, largely from external program runs
and based on input but also for overall output.

Extracted base functionality as a filename part, which itself is based
on an input <name> and optional <extra> and <suffix> args, and a file
extension part, which is largely based on the type of file and an
optional formatting arg.

May still want to improve upon this in future, for example, allowing
more than one entry into <extra>, since join() takes an iterable anyway.