Finished second round of testing for files and filter.
Finished testing for all modules in 'files' and 'filter' directories.
Note: realized that using tempfile.TemporaryDirectory() only worked in
some cases because of the calling context of the function in which the
tmpdir was defined.

A more general solution is to use tempfile.mkdtemp(), add the created
directory to scrollpy.tmps_to_remove, and then make sure to call
shutil.rmtree from either the cleanup function in __main__ or from a
tear down method in the testcase.